% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiguresAndTables.R
\name{analysePsDistributions}
\alias{analysePsDistributions}
\title{Analyse propensity score distributions}
\usage{
analysePsDistributions(workFolder)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Analyse propensity score distributions
}
\details{
This function plots all propensity score distributions, and computes AUC
and equipoise for every exposure pair.
}
