% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{calibrateEstimatesAndPvalues}
\alias{calibrateEstimatesAndPvalues}
\title{Created calibrated confidence intervals, estimates, and p-values.}
\usage{
calibrateEstimatesAndPvalues(workFolder)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}
}
\description{
Created calibrated confidence intervals, estimates, and p-values.
}
