% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Large-Scale Population-Level Evidence Generation study}
\usage{
execute(connectionDetails, cdmDatabaseSchema, oracleTempSchema,
  workDatabaseSchema, studyCohortTable, exposureCohortSummaryTable, workFolder,
  maxCores, createCohorts = TRUE, fetchAllDataFromServer = TRUE,
  injectSignals = TRUE, generateAllCohortMethodDataObjects = TRUE,
  runCohortMethod = TRUE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{exposureCohortSummaryTable}{The name of the table that will be created in the work database schema.
This table will hold summary data on the  exposure cohorts used in this
study.}

\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{createCohorts}{Create the studyCohortTable and exposureCohortSummaryTable tables with the exposure and outcome cohorts?}

\item{fetchAllDataFromServer}{Fetch all relevant data from the server?}

\item{injectSignals}{Inject signals to create synthetic controls?}

\item{generateAllCohortMethodDataObjects}{Create the cohortMethodData objects from the fetched data and injected signals?}

\item{runCohortMethod}{Run the CohortMethod package to produce the outcome models.}
}
\description{
Execute OHDSI Large-Scale Population-Level Evidence Generation study
}
\details{
This function executes the OHDSI Large-Scale Population-Level Evidence Generation study.
}
