% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCmDataObjectBuilding.R
\name{fetchAllDataFromServer}
\alias{fetchAllDataFromServer}
\title{Fetch all data on the cohorts for analysis}
\usage{
fetchAllDataFromServer(connectionDetails, cdmDatabaseSchema, workDatabaseSchema,
  studyCohortTable = "ohdsi_cohorts", oracleTempSchema, workFolder)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the study cohort table  in the work database schema.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{exposureCohortSummaryTable}{The name of the exposure summary table in the work database schema.}
}
\description{
Fetch all data on the cohorts for analysis
}
\details{
This function will create covariates and fetch outcomes and person information from the server.
}
