% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterByExposureCohortsSize.R
\name{filterByExposureCohortsSize}
\alias{filterByExposureCohortsSize}
\title{Filter exposure pairs by size of the cohorts}
\usage{
filterByExposureCohortsSize(workFolder, minCohortsSize = 2500)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{minCohortsSize}{Minimum number of people that have to be in each cohort to keep a pair of
cohorts.}
}
\description{
Filter exposure pairs by size of the cohorts
}
