% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{runCohortMethod}
\alias{runCohortMethod}
\title{Run the cohort method package}
\usage{
runCohortMethod(workFolder, maxCores = 4)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Run the cohort method package
}
\details{
Runs the cohort method package to produce propensity scores and outcome models.
}
